+++ tests/test_getitem_roundtrip.rs
@@ -50,7 +50,7 @@ fn run_getitem_from_c_compressed(
     doshuffle: i32,
 ) {
     let buffer_size = type_size * num_elements;
-    let dest_size = buffer_size + BLOSC2_MAX_OVERHEAD;
+    let dest_size = buffer_size + BLOSC2_MAX_OVERHEAD as usize;
 
     let mut original = vec![0u8; buffer_size];
     fill_seq(&mut original);
@@ -108,7 +108,7 @@ fn run_getitem_from_blusc_compressed(
     doshuffle: i32,
 ) {
     let buffer_size = type_size * num_elements;
-    let dest_size = buffer_size + BLOSC2_MAX_OVERHEAD;
+    let dest_size = buffer_size + BLOSC2_MAX_OVERHEAD as usize;
 
     let mut original = vec![0u8; buffer_size];
     fill_seq(&mut original);
@@ -146,7 +146,7 @@ fn run_getitem_from_blusc_compressed(
 /// Compress with C, then use blusc getitem to extract a subset of elements.
 fn run_getitem_subset(type_size: usize, num_elements: usize, clevel: i32, doshuffle: i32) {
     let buffer_size = type_size * num_elements;
-    let dest_size = buffer_size + BLOSC2_MAX_OVERHEAD;
+    let dest_size = buffer_size + BLOSC2_MAX_OVERHEAD as usize;
 
     let mut original = vec![0u8; buffer_size];
     fill_seq(&mut original);
